+++ src/gpu/effects/GrTextureDomain.cpp
@@ -111,7 +111,8 @@ void GrTextureDomain::GLDomain::sampleTexture(GrGLSLShaderBuilder* builder,
                                               const char* inModulateColor) {
     auto appendTextureSample = [&sampler, inModulateColor, builder](const char* coord) {
         builder->codeAppend("half4 textureColor = ");
-        builder->appendTextureLookupAndModulate(inModulateColor, sampler, coord);
+        builder->appendTextureLookupAndBlend(inModulateColor, SkBlendMode::kModulate, sampler,
+                                             coord);
         builder->codeAppend(";");
         return SkString("textureColor");
     };
@@ -607,9 +608,7 @@ GR_DEFINE_FRAGMENT_PROCESSOR_TEST(GrDeviceSpaceTextureDecalFragmentProcessor);
 #if GR_TEST_UTILS
 std::unique_ptr<GrFragmentProcessor> GrDeviceSpaceTextureDecalFragmentProcessor::TestCreate(
         GrProcessorTestData* d) {
-    int texIdx = d->fRandom->nextBool() ? GrProcessorUnitTest::kSkiaPMTextureIdx
-                                        : GrProcessorUnitTest::kAlphaTextureIdx;
-    sk_sp<GrTextureProxy> proxy = d->textureProxy(texIdx);
+    auto [proxy, at, ct] = d->randomProxy();
     SkIRect subset;
     subset.fLeft = d->fRandom->nextULessThan(proxy->width() - 1);
     subset.fRight = d->fRandom->nextRangeU(subset.fLeft, proxy->width());